requirements.txt
Django>=1.6
South
Pillow>=2.2.1
django-filer>=0.9.5
django-mptt>=0.5.2
django-sekizai>=0.7
django-widget-tweaks>=1.3
djangorestframework>=2.3.10
easy-thumbnails>=1.4
html5lib>=0.99
psycopg2>=2.5.1
wsgiref>=0.1.2
